requirements.txt
fastapi==0.136.0
uvicorn[standard]==0.44.0
sqlalchemy==2.0.49
pydantic[email]==2.13.2
pydantic-settings==2.13.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.26
python-dotenv==1.2.2
pytest==7.4.0
httpx==0.25.0